requirements.txt
# Required
pyrogram
TgCrypto
python-dotenv
hachoir
ffmpeg-python
beautifulsoup4==4.9.1
bs4==0.0.1

# DDL
pySmartDL
requests

# Database
dnspython
motor

# Others
psutil
js2py
cfscrape
lk21
lxml
httpx

# Google Drive
google-api-core
google-auth
oauth2client<4.0.0
google~=3.0.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
httplib2==0.18.0
aiohttp
# env
python-dotenv

